requirements.txt
GitPython
markdown
sqlite-utils>=3.2
